--- src/dragonfly_reporter/__main__.py
@@ -1,5 +1,0 @@
-import uvicorn
-
-from dragonfly_reporter.app import app
-
-uvicorn.run(app)
