+++ src/reporter/__main__.py
@@ -1,0 +1,5 @@
+import uvicorn
+
+from reporter.app import app
+
+uvicorn.run(app)
